+++ R/tissot-package.R
@@ -1,0 +1,17 @@
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL
+
+#' world coastline
+#'
+#' A modified matrix version of data from the maps package.
+#'
+#' Basically longitudes have been smooshed to 'abs(lon) < 180'
+#' @docType data
+#' @name world
+'world'