+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Improved ability to build more than one tissot and to easily plot multiple indicatrixes. 
 
+* New function `tissot_get_proj()` with no arguments, just gets the last value of a plot projection
+registred by plot.indicatrixes. 
+
 
 # tissot 0.0.1